requirements.txt
asgiref==3.11.0
Django==6.0
django-filter==25.2
django-phonenumber-field==8.4.0
django-phonenumbers==1.0.1
django-resized==1.0.3
django-restframework==0.0.1
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.21
pillow==12.0.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
tzdata==2025.3
uritemplate==4.2.0